Labb4 Assignment 1 and 2
First assignment we make functions to control an external EEPROM, this is done by sending control bit with READ or WRITE bit (see idc.c) and the address in EEPROM where we want to store in.

Assignment 2 takes advantage of these functions in order to write bytes to (multiple) addressess and reading from said addresses.
Check comments in idc.c for more information,

In main.c is a demostration how its done.